Separated the trcuk class into manager and truck
