+++ DAN_XXXVII_Kristina_Garcia_Francisco/DAN_XXXVII_Kristina_Garcia_Francisco.csproj
@@ -43,6 +43,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Manager.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ReadWriteFile.cs" />
